+++ src/database/poo.sql
@@ -1,4 +1,4 @@
--- Active: 1675168195024@@127.0.0.1@3306
+-- Active: 1683066430375@@127.0.0.1@3306
 CREATE TABLE users (
     id TEXT PRIMARY KEY UNIQUE NOT NULL,
     name TEXT NOT NULL,
